generation pub is per user

+++ src/main/java/it/polimi/streamreasoning/xbench/model/DiscussionLeaderModel.java
@@ -3,5 +3,5 @@ package it.polimi.streamreasoning.xbench.model;
 /** information of an RA instance */
 public class DiscussionLeaderModel {
     public int type = 0;
-    public int id = 0;
+    public long id = 0;
 }
